Wait for ExecCommand before colorizing
ExecCommand uses a short timeout internally to append the data to the view. If this timeout isn't fired fast enough the regions can't be colored. A simple solution is to also add a timeout to the colorization.

This could still be improved a little by checking whether the regions are within the view size, and only use a timeout if they are not. That would remove some rare flickering introduced by the timeout.